feat: adds a healthchecks to compose file
Adds a extremely basic healthcheck endpoint to the Caddy container (localhost:8081/health). It's on a sepatate port so
it can be used without exposing it to the world.

Adds healthcheck config and gates to the compose file. This doesn't really buy much as the services are too simple to
really fail but having them set up puts us in a more robust position.