feat: switches to cusomised Caddy container
Using the populate container and volumes was proving fragile. Instead this change adds a stage to the Dockerfile that
creates a customised Caddy container with the static assets and Caddyfile baked in.
Adds a separate npm script to build the caddy container (stop at a different target and give it a different name)
Removes populate container.
Makes db container optional with `internal_db` profile.